+ Updated version to 5.12.4.2 + Fixed some net opcode handling (thanks to purple for all the help)
git-svn-id: https://svn.code.sf.net/p/seq/svn/showeq/trunk@705 19a12193-7728-0410-8f54-e9f61ebc0c9f

+++ src/packetformat.cpp
@@ -93,7 +93,7 @@ EQProtocolPacket& EQProtocolPacket::operator=(const EQProtocolPacket& packet)
   return *this;
 }
 
-void EQProtocolPacket::init(uint8_t* packet, uint16_t length, 
+void EQProtocolPacket::init(uint8_t* packet, uint32_t length, 
   bool copy, bool subpacket)
 {
   m_subpacket = subpacket;
